pyproject.toml
[project]
name = "filerenamer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colorama>=0.4.6",
    "google-genai>=1.32.0",
    "python-dotenv>=1.1.1",
]
